backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Pillow>=10.0.0
numpy>=1.24.0
onnxruntime>=1.17.0
python-multipart==0.0.6
python-dotenv==1.0.0
